Moved movements table component to its own file
